Add files via upload
WordWiz is a simple word guessing game built in Python. The game selects a random word from a dictionary, provides a hint about the word's meaning, and allows the player to guess the word. The player has a limited number of turns to guess the word correctly.

Features
Random Word Selection: The game randomly selects a word from a dictionary with associated hints.
Limited Turns: The player has a limited number of turns to guess the word.
User Interaction: The player can input their guess and interact with the game.
How to Play
1- The game starts by asking the player to enter their name.
2- The player is greeted and asked if they want to start the game.
3- If the player chooses to play (by entering 'y'), the game proceeds.
4- The game selects a random word from the dictionary along with an associated hint.
5- The player is informed about the total number of alphabets in the selected word and is prompted to guess the word.
6- The player can input their guess (case-insensitive).
7- If the player guesses the word correctly, they win the game, and the game ends.
8- If the player's guess is incorrect, the game informs them and deducts one turn.
9- The game continues until the player guesses the word correctly or runs out of turns.
10- If the player runs out of turns without guessing the word, the game reveals the correct word and ends.
11- After the game, the player has the option to see the game's results (total alphabets, remaining turns, and failed attempts).